dom/media/test/test_loop_clock.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Test currentTime when looping</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
  <script type="text/javascript" src="manifest.js"></script>
</head>
<body>
<pre id="test">
<script class="testbody" type="text/javascript">
SimpleTest.waitForExplicitFinish();

var v = document.createElement('video');
v.src = "gizmo-short.mp4";
v.seekedCount = 0;
v.load();
v.loop = true;
v.controls = true;
v.muted = true;
v.addEventListener("loadedmetadata", function() {
  v.initialDuration = v.duration;
  v.play();
  v.intervalHandle = setInterval(function() {
    if (v.currentTime > v.initialDuration) {
      ok(false, `currentTime should never exceed the initial duration
         (iteration ${v.seekedCount})) (${v.currentTime} <= ${v.initialDuration})`);
      clearInterval(v.intervalHandle);
      removeNodeAndSource(v);
      SimpleTest.finish();
    }
  }, 1);
});

v.addEventListener("seeked", function (e) {
  e.target.seekedCount += 1;
  var iterations = 50;
  if (e.target.seekedCount == iterations) {
    clearInterval(v.intervalHandle);
    removeNodeAndSource(v);
    SimpleTest.finish();
  }
});
document.body.appendChild(v);

</script>
</pre>
</body>
</html>
